Fixed compiler warnings

+++ include/constants.hpp
@@ -5,8 +5,8 @@
 #define MD_E_DIGITS 2.71828182845904523
 
 #define MD_DEFINE_CONSTANT(name, constant)                                     \
-	constexpr float name##_f = constant;                                       \
-	constexpr double name##_d = constant;
+	constexpr float name##_f = float(constant);                                       \
+	constexpr double name##_d = double(constant);
 
 namespace md_math {
 MD_DEFINE_CONSTANT(pi, MD_PI_DIGITS)
